refactored pharmacy
