+++ components/pharmacy/stock/MedicationModal.js
